Expansion of the package
- Everything is in separate classes/files
- the test suite is the same but was moved out of the scholarly folder
- implemented a navigator class, that is kind of a "browser" to better distribute tasks among classes
- updated action to take in account modifications.
- Since this is a major relesease I its own branch
- Made ToR more platform robust.
