+++ scholarly/__init__.py
@@ -1,4 +1,3 @@
-from .scholarly import Author, Publication
-from .scholarly import search_author, search_keyword, search_pubs_query
-from .scholarly import search_pubs_custom_url, search_author_custom_url
-from .scholarly import use_proxy, use_tor
+from ._scholarly import _Scholarly
+
+scholarly = _Scholarly()
