+++ scholarly/publication.py
@@ -1,0 +1,171 @@
+import re
+import bibtexparser
+import arrow
+import requests
+
+_SCHOLARPUBRE = r'cites=([\w-]*)'
+_CITATIONPUB = '/citations?hl=en&view_op=view_citation&citation_for_view={0}'
+_SCHOLARPUB = '/scholar?hl=en&oi=bibs&cites={0}'
+_CITATIONPUBRE = r'citation_for_view=([\w-]*:[\w-]*)'
+
+
+class _SearchScholarIterator(object):
+    """Iterator that returns Publication objects from the search page
+    I have removed all logging from here for simplicity. -V
+    """
+
+    def __init__(self, nav, url:str):
+        self._url = url
+        self._nav = nav
+        self._load_url(url)
+
+    def _load_url(self, url:str):
+        #this is temporary until setup json file
+        self._soup = self._nav._get_soup(url)
+        self._pos = 0
+        self._rows = self._soup.find_all('div', class_='gs_r gs_or gs_scl')
+
+    # Iterator protocol
+
+    def __iter__(self):
+        return self
+
+    def __next__(self):
+        if self._pos < len(self._rows):
+            row = self._rows[self._pos]
+            self._pos += 1
+            return Publication(self._nav, row, 'scholar')
+        elif self._soup.find(class_='gs_ico gs_ico_nav_next'):
+            url = self._soup.find(
+                class_='gs_ico gs_ico_nav_next').parent['href']
+            self._load_url(url)
+            return self.__next__()
+        else:
+            raise StopIteration
+
+    # Pickle protocol
+    def __getstate__(self):
+        return {'url': self._url, 'pos': self._pos}
+
+    def __setstate__(self, state):
+        #this needs validation -V
+        self._load_url(state['url'])
+        self._pos = state['pos']
+
+class Publication(object):
+    """Returns an object for a single publication"""
+
+    def __init__(self, nav, __data, pubtype=None):
+        self.nav = nav
+        self.bib = dict()
+        self.source = pubtype
+        if self.source == 'citations':
+            self.bib['title'] = __data.find('a', class_='gsc_a_at').text
+            self.id_citations = re.findall(_CITATIONPUBRE, __data.find(
+                'a', class_='gsc_a_at')['data-href'])[0]
+            citedby = __data.find(class_='gsc_a_ac')
+            if citedby and not (citedby.text.isspace() or citedby.text == ''):
+                self.citedby = int(citedby.text)
+            year = __data.find(class_='gsc_a_h')
+            if year and year.text and not year.text.isspace() and len(year.text) > 0:
+                self.bib['year'] = int(year.text)
+        elif self.source == 'scholar':
+            databox = __data.find('div', class_='gs_ri')
+            title = databox.find('h3', class_='gs_rt')
+            if title.find('span', class_='gs_ctu'):  # A citation
+                title.span.extract()
+            elif title.find('span', class_='gs_ctc'):  # A book or PDF
+                title.span.extract()
+            self.bib['title'] = title.text.strip()
+            if title.find('a'):
+                self.bib['url'] = title.find('a')['href']
+            authorinfo = databox.find('div', class_='gs_a')
+            self.bib['author'] = ' and '.join(
+                [i.strip() for i in authorinfo.text.split(' - ')[0].split(',')])
+            try:
+                self.bib['venue'], self.bib['year'] = authorinfo.text.split(
+                    ' - ')[1].split(',')
+            except Exception as e:
+                self.bib['venue'], self.bib['year'] = 'NA', 'NA'
+            if databox.find('div', class_='gs_rs'):
+                self.bib['abstract'] = databox.find('div', class_='gs_rs').text
+                if self.bib['abstract'][0:8].lower() == 'abstract':
+                    self.bib['abstract'] = self.bib['abstract'][9:].strip()
+            lowerlinks = databox.find('div', class_='gs_fl').find_all('a')
+            for link in lowerlinks:
+                if 'Import into BibTeX' in link.text:
+                    self.url_scholarbib = link['href']
+                if 'Cited by' in link.text:
+                    self.citedby = int(re.findall(r'\d+', link.text)[0])
+                    self.id_scholarcitedby = re.findall(
+                        _SCHOLARPUBRE, link['href'])[0]
+            if __data.find('div', class_='gs_ggs gs_fl'):
+                self.bib['eprint'] = __data.find(
+                    'div', class_='gs_ggs gs_fl').a['href']
+        self._filled = False
+
+    def fill(self):
+        """Populate the Publication with information from its profile"""
+        if self.source == 'citations':
+            url = _CITATIONPUB.format(self.id_citations)
+            soup = self.nav._get_soup(url)
+            self.bib['title'] = soup.find('div', id='gsc_vcd_title').text
+            if soup.find('a', class_='gsc_vcd_title_link'):
+                self.bib['url'] = soup.find(
+                    'a', class_='gsc_vcd_title_link')['href']
+            for item in soup.find_all('div', class_='gs_scl'):
+                key = item.find(class_='gsc_vcd_field').text
+                val = item.find(class_='gsc_vcd_value')
+                if key == 'Authors':
+                    self.bib['author'] = ' and '.join(
+                        [i.strip() for i in val.text.split(',')])
+                elif key == 'Journal':
+                    self.bib['journal'] = val.text
+                elif key == 'Volume':
+                    self.bib['volume'] = val.text
+                elif key == 'Issue':
+                    self.bib['number'] = val.text
+                elif key == 'Pages':
+                    self.bib['pages'] = val.text
+                elif key == 'Publisher':
+                    self.bib['publisher'] = val.text
+                elif key == 'Publication date':
+                    self.bib['year'] = arrow.get(val.text).year
+                elif key == 'Description':
+                    if val.text[0:8].lower() == 'abstract':
+                        val = val.text[9:].strip()
+                    self.bib['abstract'] = val
+                elif key == 'Total citations':
+                    self.id_scholarcitedby = re.findall(
+                        _SCHOLARPUBRE, val.a['href'])[0]
+
+            # number of citation per year
+            years = [int(y.text) for y in soup.find_all(class_='gsc_vcd_g_t')]
+            cites = [int(c.text) for c in soup.find_all(class_='gsc_vcd_g_al')]
+            self.cites_per_year = dict(zip(years, cites))
+
+            if soup.find('div', class_='gsc_vcd_title_ggi'):
+                self.bib['eprint'] = soup.find(
+                    'div', class_='gsc_vcd_title_ggi').a['href']
+            self._filled = True
+        elif self.source == 'scholar':
+            bibtex = self.nav._get_page(self.url_scholarbib)
+            self.bib.update(bibtexparser.loads(bibtex).entries[0])
+            self._filled = True
+        return self
+
+    def get_citedby(self) -> _SearchScholarIterator or list:
+        """Searches GScholar for other articles that cite this Publication and
+        returns a Publication generator.
+        """
+        if not hasattr(self, 'id_scholarcitedby'):
+            self.fill()
+        if hasattr(self, 'id_scholarcitedby'):
+            url = _SCHOLARPUB.format(
+                requests.utils.quote(self.id_scholarcitedby))
+            return _SearchScholarIterator(self.nav, url)
+        else:
+            return []
+
+    def __str__(self):
+        return pprint.pformat(self.__dict__)
